requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.1.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
slowapi==0.1.9
email-validator==2.1.0
psycopg2-binary==2.9.9

# Scraping e automação
selenium==4.15.2
webdriver-manager==4.0.1

